# packages/core-schemas/openapi.yaml
openapi: 3.1.0
info:
  title: DreamRight API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamRight Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}/characters/{characterId}/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/characterId'

    get:
      tags: [Characters]
      summary: Get character assets metadata
      operationId: getCharacterAssets
      responses:
        '200':
          description: Character assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterAssets'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Characters]
      summary: Generate character portrait
      description: |
        Generates a character portrait image using AI.
        The portrait is used as a reference for panel generation.
        This is an async operation.
      operationId: createCharacterAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Asset generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Asset already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExistsError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Characters]
      summary: Regenerate character portrait
      description: Force regenerate the character portrait, replacing existing asset.
      operationId: replaceCharacterAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Asset regeneration started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/characters/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'

    post:
      tags: [Characters]
      summary: Generate all missing character assets
      description: Batch generate portraits for all characters without assets.
      operationId: createAllCharacterAssets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterAssetRequest'
      responses:
        '202':
          description: Batch generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

  # Location endpoints
...
  /projects/{projectId}/locations/{locationId}/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/locationId'

    get:
      tags: [Locations]
      summary: Get location assets metadata
      operationId: getLocationAssets
      responses:
        '200':
          description: Location assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationAssets'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Locations]
      summary: Generate location reference
      description: |
        Generates a location reference image using AI.
        This is an async operation.
      operationId: createLocationAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationAssetRequest'
      responses:
        '202':
          description: Asset generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Asset already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExistsError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Locations]
      summary: Regenerate location reference
      description: Force regenerate the location reference, replacing existing asset.
      operationId: replaceLocationAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationAssetRequest'
      responses:
        '202':
          description: Asset regeneration started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/locations/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'

    post:
      tags: [Locations]
      summary: Generate all missing location assets
      description: Batch generate references for all locations without assets.
      operationId: createAllLocationAssets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationAssetRequest'
      responses:
        '202':
          description: Batch generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

  # Chapter endpoints - Resource-oriented with stable IDs
...
  /projects/{projectId}/assets:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Assets]
      summary: List all assets
      description: List all generated asset files for a project
      operationId: listAssets
      parameters:
        - name: type
          in: query
          description: Filter by asset type
          schema:
            type: string
            enum: [character, location, panel]
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAssetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/assets/{assetType}/{assetId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: assetType
        in: path
        required: true
        description: Type of asset (characters, locations, panels)
        schema:
          type: string
          enum: [characters, locations, panels]
      - name: assetId
        in: path
        required: true
        description: Asset identifier (e.g., character slug, panel path)
        schema:
          type: string

    get:
      tags: [Assets]
      summary: Get asset file
      description: Retrieve a generated asset file (image)
      operationId: getAssetFile
      responses:
        '200':
          description: Asset file
          content:
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/assets/{assetType}/{assetId}/metadata:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: assetType
        in: path
        required: true
        schema:
          type: string
          enum: [characters, locations, panels]
      - name: assetId
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Assets]
      summary: Get asset metadata
      description: Retrieve generation metadata for an asset
      operationId: getAssetMetadata
      responses:
        '200':
          description: Asset metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

  # Job endpoints for async operations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
    characterId:
      name: characterId
      in: path
      required: true
      description: Character ID
      schema:
        type: string
    locationId:
      name: locationId
      in: path
      required: true
      description: Location ID
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ErrorCode:
      type: string
      enum:
        - bad_request
        - validation_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - dependency_not_met
        - asset_exists
        - rate_limited
        - internal_error
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AssetExistsError:
      type: object
      required: [code, message, asset_path]
      properties:
        code:
          type: string
          enum: [asset_exists]
        message:
          type: string
        asset_path:
          type: string
          description: Path to existing asset
        use_put_to_replace:
          type: boolean
          default: true

    # Pagination wrapper
    PaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Paginated lists
...
    PaginatedAssetList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssetInfo'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Job schemas
...
    JobResponse:
      type: object
      required: [job_id, status, poll_url]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
          description: URL to poll for job status
    BatchJobResponse:
      type: object
      required: [job_id, status, poll_url, total_items]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
        total_items:
          type: integer
          description: Number of items to process

    # Project schemas
...
    CharacterAssets:
      type: object
      properties:
        portrait:
          type: string
          nullable: true
          description: Path to portrait image
        portrait_url:
          type: string
          format: uri
          nullable: true
          description: URL to download portrait
        three_view:
          type: object
          properties:
            front:
              type: string
              nullable: true
            side:
              type: string
              nullable: true
            back:
              type: string
              nullable: true
...
    CreateCharacterAssetRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation
        portrait_only:
          type: boolean
          default: true
          description: Generate only portrait (not three-view)

    # Location schemas
    LocationAssets:
      type: object
      properties:
        reference:
          type: string
          nullable: true
          description: Path to reference image
        reference_url:
          type: string
          format: uri
          nullable: true
          description: URL to download reference
...
    CreateLocationAssetRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation

    # Chapter schemas
...
    AssetInfo:
      type: object
      required: [id, type, path, created_at]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [character, location, panel]
        path:
          type: string
        url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        metadata_url:
          type: string
          format: uri
    AssetMetadata:
      type: object
      required: [type, generated_at]
      properties:
        type:
          type: string
        generated_at:
          type: string
          format: date-time
        prompt:
          type: string
          description: The prompt used for generation
        parameters:
          type: object
          description: Generation parameters
          properties:
            style:
              type: string
            resolution:
              type: string
            aspect_ratio:
              type: string
            model:
              type: string
        references:
          type: array
          description: Reference images used
          items:
            type: object
            properties:
              path:
                type: string
              role:
                type: string